REmove numpy 1.24.0 deprecation warning

+++ src/silx/math/test/test_combo.py
@@ -1,5 +1,5 @@
 # /*##########################################################################
-# Copyright (C) 2016-2020 European Synchrotron Radiation Facility
+# Copyright (C) 2016-2022 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -137,7 +137,7 @@ class TestMinMax(ParametricTestCase):
                                       min_positive=min_positive,
                                       data=name):
                         data = numpy.arange(
-                            start, start + step * size, step, dtype=dtype)
+                            start, start + step * size, step).astype(dtype)
 
                         self._test_min_max(data, min_positive)
